Fold AbstractTransport into TcpTransport
It's hard to see what it's providing which is useful outside
TcpTransport at the moment. Perhaps as we add further transports, that
will become clear and we can pull out commonality then.
